@@ -11,13 +11,13 @@
   "author": "Aaron Smith (asmith@aims.gov.au)",
   "license": "CC-BY-4.0",
   "devDependencies": {
-    "babel-core": "^6.26.0",
+    "babel-core": "^6.26.3",
     "babel-loader": "^7.1.2",
     "babel-preset-env": "^1.6.1",
     "clean-webpack-plugin": "^0.1.17",
-    "eslint": "^4.15.0",
-    "eslint-config-airbnb-base": "^12.1.0",
-    "eslint-plugin-import": "^2.8.0",
+    "eslint": "^4.19.1",
+    "eslint-config-aimsks-base": "git+https://github.com/aims-ks/eslint-config-aimsks-base.git",
+    "eslint-plugin-import": "^2.14.0",
     "webpack": "^3.10.0"
   },
   "dependencies": {
