@@ -1,10 +1,9 @@
 /**
  * A Model that fires an event when the underlying data is changed.
  */
-const EventEmitter = require('events').EventEmitter;
+import EventEmitter from 'events';
 
-module.exports = class BasicModel extends EventEmitter {
-
+export default class BasicModel extends EventEmitter {
   /**
    * Constructor to define the cached data and the event(s) that can be fired by this class.
    */
@@ -39,5 +38,4 @@ module.exports = class BasicModel extends EventEmitter {
   getData() {
     return this._cachedData;
   }
-
-};
+}
