@@ -2,10 +2,9 @@
  * A map-based Model that fires an event when the underlying data is changed. This model expects
  * each entry to have a unique `id`.
  */
-const EventEmitter = require('events').EventEmitter;
+import EventEmitter from 'events';
 
-module.exports = class MapModel extends EventEmitter {
-
+export default class MapModel extends EventEmitter {
   /**
    * Constructor to define the cached data and the event(s) that can be fired by this class.
    */
@@ -30,16 +29,15 @@ module.exports = class MapModel extends EventEmitter {
    * {@link #_cachedData} is modified, a change event will be emitted.
    */
   add(data, forceEmit) {
-
     // Flag to determine if cachedData is updated.
     let isUpdated = false;
 
     if (Array.isArray(data)) {
       // data is an array.
-      for (entry of data) {
+      data.forEach((entry) => {
         this._cachedData[entry.id] = entry;
         isUpdated = true;
-      }
+      });
     } else {
       // Treat data as an object.
       Object.keys(data).forEach((key) => {
@@ -52,7 +50,6 @@ module.exports = class MapModel extends EventEmitter {
     if (isUpdated || forceEmit) {
       this.emit(this.EVENT_DATA_CHANGED, this);
     }
-
   }
 
   /**
@@ -68,29 +65,27 @@ module.exports = class MapModel extends EventEmitter {
    * modified, a change event will be emitted.
    */
   remove(data) {
-
     // Flag to determine if cachedData is updated.
     let isUpdated = false;
 
     if (Array.isArray(data)) {
       // data is an array.
-      for (entry of data) {
+      data.forEach((entry) => {
         delete this._cachedData[entry.id];
         isUpdated = true;
-      }
-    } else {
-      // Treat data as a string.
-      if (this._cachedData[data]) {
-        delete this._cachedData[data];
-        isUpdated = true;
-      }
+      });
+    }
+
+    // Check if data is a string that identifies a valid id.
+    if ((typeof data === 'string') && this._cachedData[data]) {
+      delete this._cachedData[data];
+      isUpdated = true;
     }
 
     // Emit a change event.
     if (isUpdated) {
       this.emit(this.EVENT_DATA_CHANGED, this);
     }
-
   }
 
   /**
@@ -104,5 +99,4 @@ module.exports = class MapModel extends EventEmitter {
     this._cachedData = {};
     this.add(data, forceEmit);
   }
-
-};
+}
